+++ types/protocol-mapping.d.ts
@@ -556,6 +556,8 @@ export namespace ProtocolMapping {
          * The following parameters are included in all events.
          */
         'Storage.sharedStorageAccessed': [Protocol.Storage.SharedStorageAccessedEvent];
+        'Storage.storageBucketCreatedOrUpdated': [Protocol.Storage.StorageBucketCreatedOrUpdatedEvent];
+        'Storage.storageBucketDeleted': [Protocol.Storage.StorageBucketDeletedEvent];
         /**
          * Issued when attached to target because of auto-attach or `attachToTarget` command.
          */
@@ -4202,6 +4204,20 @@ export namespace ProtocolMapping {
             returnType: void;
         };
         /**
+         * Set tracking for a storage key's buckets.
+         */
+        'Storage.setStorageBucketTracking': {
+            paramsType: [Protocol.Storage.SetStorageBucketTrackingRequest];
+            returnType: void;
+        };
+        /**
+         * Deletes the Storage Bucket with the given storage key and bucket name.
+         */
+        'Storage.deleteStorageBucket': {
+            paramsType: [Protocol.Storage.DeleteStorageBucketRequest];
+            returnType: void;
+        };
+        /**
          * Returns information about the system.
          */
         'SystemInfo.getInfo': {
@@ -4664,6 +4680,14 @@ export namespace ProtocolMapping {
         };
         'FedCm.dismissDialog': {
             paramsType: [Protocol.FedCm.DismissDialogRequest];
+            returnType: void;
+        };
+        /**
+         * Resets the cooldown time, if any, to allow the next FedCM call to show
+         * a dialog even if one was recently dismissed by the user.
+         */
+        'FedCm.resetCooldown': {
+            paramsType: [];
             returnType: void;
         };
     }